feat(eslint-plugin): [no-shadow] add shadowed variable location to the error message
