+++ packages/eslint-plugin/src/rules/no-shadow.ts
@@ -12,7 +12,7 @@ import {
 } from '@typescript-eslint/scope-manager';
 import * as util from '../util';
 
-type MessageIds = 'noShadow';
+type MessageIds = 'noShadow' | 'noShadowGlobal';
 type Options = [
   {
     allow?: string[];
@@ -64,7 +64,9 @@ export default util.createRule<Options, MessageIds>({
       },
     ],
     messages: {
-      noShadow: "'{{name}}' is already declared in the upper scope.",
+      noShadow:
+        "'{{name}}' is already declared in the upper scope on line {{shadowedLine}} column {{shadowedColumn}}.",
+      noShadowGlobal: "'{{name}}' is already a global variable.",
     },
   },
   defaultOptions: [
@@ -518,6 +520,28 @@ export default util.createRule<Options, MessageIds>({
     }
 
     /**
+     * Get declared line and column of a variable.
+     * @param  variable The variable to get.
+     * @returns The declared line and column of the variable.
+     */
+    function getDeclaredLocation(
+      variable: TSESLint.Scope.Variable,
+    ): { global: true } | { global: false; line: number; column: number } {
+      const identifier = variable.identifiers[0];
+      if (identifier) {
+        return {
+          global: false,
+          line: identifier.loc.start.line,
+          column: identifier.loc.start.column + 1,
+        };
+      } else {
+        return {
+          global: true,
+        };
+      }
+    }
+
+    /**
      * Checks the current context for shadowed variables.
      * @param {Scope} scope Fixme
      */
@@ -595,12 +619,26 @@ export default util.createRule<Options, MessageIds>({
           ) &&
           !(options.hoist !== 'all' && isInTdz(variable, shadowed))
         ) {
+          const location = getDeclaredLocation(shadowed);
+          const report = location.global
+            ? ({
+                messageId: 'noShadowGlobal',
+                data: {
+                  name: variable.name,
+                },
+              } as const)
+            : ({
+                messageId: 'noShadow',
+                data: {
+                  name: variable.name,
+                  shadowedLine: location.line,
+                  shadowedColumn: location.column,
+                },
+              } as const);
+
           context.report({
             node: variable.identifiers[0],
-            messageId: 'noShadow',
-            data: {
-              name: variable.name,
-            },
+            ...report,
           });
         }
       }
